requirements.txt
# 核心区块链依赖
web3==7.14.0

# 测试框架+报告工具
pytest==9.0.1
allure-pytest==2.15.2